Extract CompressionConfigUtil
